requirements.txt
gym==0.18.3
numpy==1.21.0
pandas==1.2.5
matplotlib==3.4.2
attrs==21.2.0
dynaconf==3.1.4
binance-connector==1.4.0